rust: replace pushd usage in rust do_install
In dizzy, the following error was encountered when running do_install
on the rust recipe:

    pushd: not found

According to the bitbake user manual, /bin/sh is used and bashisms
(like pushd) should be avoided.  In my case, /bin/sh turns out to
be zsh (with no extra configuration, so pushd is not present).

Signed-off-by: Paul Osborne <osbpau@gmail.com>