Don't build any rlibs
I've seen this error from the compiler:
    found possibly newer version of crate `regex_syntax` which
    `env_logger` depends on

This is probably because the rlib hashes are not deterministic, but I
can't reproduce non-deterministic hashes locally.

There have been lots of other types of errors caused by mixing rlibs and
dylibs in the same dependency chain, so let's try using only dylibs
until that's stable.