classes/cargo: implement a local sysroot crate index
Cargo packages publish their crate tarball and metadata to the sysroot.
Clients create a local git repo of the metadata directory to use as a
sysroot that points to the sysroot as the download URL. This allows
bitbake to manage crate source downloading and availability. Crates
should only see other crates that have been built by bitbake.

Signed-off-by: Tyler Hall <tylerwhall@gmail.com>