classes/cargo: allow overriding crate types
Cargo's default behavior is to only create rlibs and it is left up to
the package to change this. Shared objects are usually preferable in
OpenEmbedded since we have control over all the dependencies at once and
don't need the portability of static linking.

Create a method to override this setting uniformly.

Signed-off-by: Tyler Hall <tylerwhall@gmail.com>