fetch: fix check for local ref path
This is a fast case if the slave has a local checkout available so that
it doesn't have to do a full clone. The check unfortunately didn't take
into account all cases and failed on the new slave.

(cherry picked from commit d867f3d470bcc5548468fe3cc72786a971ffdfac)