remove recipes-core & recipes-graphics
Remove these recipes as I do not believe they are functional since the
fixes for #126 were done (removing the modifications to the compiler to
allow them to be used). #197 additionally points out that rand-rs has
not had a valid commit hash and most of these in some way depend on that
so these cannot possibly work in their current form. Removing per #102.
Fixes #102.