workflows: Remove the size reduction part from build config
As it currently looks like the sstate mirror isn't used, it might
be due to the very much modified configuration. Make it is close as
possible to the standard options to check how much from sstate will be
picked. At the same time, it'll show whether the size issue will be
still triggered.

Signed-off-by: Anatol Belski <anbelski@linux.microsoft.com>